+++ src/main/java/com/alan199921/astral/events/astraltravel/AstralTravelRendering.java
@@ -16,7 +16,7 @@ import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-@Mod.EventBusSubscriber(modid = Astral.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
+@Mod.EventBusSubscriber(modid = Astral.MOD_ID, value = Dist.CLIENT)
 public class AstralTravelRendering {
 
     @SubscribeEvent
